+++ package.json
@@ -32,8 +32,8 @@
   },
   "devDependencies": {
     "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@microsoft/api-extractor": "^7.11.2",
-    "@microsoft/tsdoc-config": "^0.13.5",
+    "@microsoft/api-extractor": "^7.12.0",
+    "@microsoft/tsdoc-config": "^0.13.6",
     "@types/aws4": "^1.5.1",
     "@types/bl": "^2.1.0",
     "@types/bson": "^4.0.2",
@@ -71,9 +71,9 @@
     "standard-version": "^7.1.0",
     "through2": "^3.0.1",
     "ts-node": "^9.0.0",
-    "typedoc": "^0.18.0",
+    "typedoc": "^0.19.2",
     "typedoc-plugin-pages": "^1.0.1",
-    "typescript": "^4.0.2",
+    "typescript": "^4.0.5",
     "typescript-cached-transpile": "^0.0.6",
     "worker-farm": "^1.5.0",
     "wtfnode": "^0.8.2",
