+++ src/index.ts
@@ -296,4 +296,3 @@ export type {
   ClientMetadataOptions
 } from './utils';
 export type { WriteConcern, W, WriteConcernOptions } from './write_concern';
-export type { MongoOptions } from './mongo_client_options';
